requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-openai
crewai
langgraph
langfuse
httpx
typing-extensions